fix: redirect contact links to data library
